Switched argument passing type for `get_counts_for`
Initially it used `*args` to get all the InstantaneousMetrics objects, but now it expects a single list. This is because generally the `parse_data` function is mapped to the return results from the database, leaving us with a list, so passing the list in directly would be cleaner than unpacking the list in the function call.